inputs from int to float

+++ src/header/perceptron.hpp
@@ -8,9 +8,9 @@ class Perceptron
 public:
     Perceptron(std::vector<double> weights, double bias, double learningRate);
 
-    int predict(const std::vector<int>& x) const;
-    void update(const std::vector<std::vector<int>>& inputs, const std::vector<int>& targets, int epochs);
-    double loss(const std::vector<std::vector<int>>& inputs, const std::vector<int>& targets) const;
+    int predict(const std::vector<float>& x) const;
+    void update(const std::vector<std::vector<float>>& inputs, const std::vector<int>& targets, int epochs);
+    double loss(const std::vector<std::vector<float>>& inputs, const std::vector<int>& targets) const;
     void __str__(int verbose) const;
     
 private:
